web: drop registration required alert
Agama already prevents users from starting the installation if any
requirements are not met and informs them at the appropriate time.
Therefore, continuously drawing the user's attention to product
registration on every page during setup is overwhelming and could be
disruptive. This commit removes the alert, as users will be properly
notified about registration, if needed, when they confirm they are ready
for installation.

Related to bsc#1253708